+++ apps/vscode/preinstall.sh
@@ -1,7 +1,7 @@
 #!/usr/bin/env bash
 
-sudo cp vscode.list /etc/apt/sources.list.d/vscode.list
-
-sudo chown root /etc/apt/sources.list.d/vscode.list
-sudo chgrp root /etc/apt/sources.list.d/vscode.list
-sudo chmod 644 /etc/apt/sources.list.d/vscode.list
+if [ ! -f "/etc/apt/sources.list.d/vscode.list" ]; then
+    sudo cp vscode.list /etc/apt/sources.list.d/vscode.list
+    sudo chown root:root /etc/apt/sources.list.d/vscode.list
+    sudo chmod 644 /etc/apt/sources.list.d/vscode.list
+fi
